tests/wpt/tests/encoding-detection/el-windows-1253-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>el windows-1253 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1253", 'Expected windows-1253');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/el-windows-1253-late.sub.html");
</script>
